requirements.txt
pandas
numpy
scikit-learn
pyarrow
fastparquet
requests
python-dotenv
hopsworks